CascadeProjects/Intune/requirements.txt
flask==3.0.0
msal==1.26.0
azure-identity==1.15.0
msgraph-sdk==1.0.0
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
